Remove reference to non-existing exportsyms.uk file .
This library used to have an exportsyms.uk file, but no longer
does. When it was removed, the reference to it in Makefile.uk was not
removed, leaving a dangling reference to a non-existing file. The
recent Unikraft build system series turned this into a build error, so
this patch fixes that by removing the line.

Signed-off-by: Felipe Huici <felipe.huici@neclab.eu>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>